清理

+++ src/idv/kuchu/mybmi/component/DataAnalyzePanel.java
@@ -29,7 +29,7 @@ public class DataAnalyzePanel extends Panel {
 
 		float v1 = weight / ((height / 100) * (height / 100));
 
-		JTextField showBMI = new JTextField(String.valueOf(((float) (int) (v1 * 100)) / 100f));// 計算
+		JTextField showBMI = new JTextField(String.valueOf(((int) (v1 * 100)) / 100f));// 計算
 		showBMI.setFont(font);
 		showBMI.setEditable(false);
 		showBMI.setBounds(200 + X, Y, 200, 64);
@@ -38,7 +38,7 @@ public class DataAnalyzePanel extends Panel {
 		labelBodyFat.setFont(font);
 		labelBodyFat.setBounds(X, 100 + Y, 164, 64);
 		JTextField showBodyFat = new JTextField(
-				String.valueOf(((float) (int) ((1.2 * v1 + 0.23 * age - 5.4 - 10.8 * gender) * 100)) / 100f));// 計算
+				String.valueOf(((int) ((1.2 * v1 + 0.23 * age - 5.4 - 10.8 * gender) * 100)) / 100f));// 計算
 		showBodyFat.setFont(font);
 		showBodyFat.setEditable(false);
 		showBodyFat.setBounds(200 + X, 100 + Y, 200, 64);
@@ -54,7 +54,7 @@ public class DataAnalyzePanel extends Panel {
 			v2 = (height - 70) * 0.6f;
 		}
 
-		JTextField showIdealweight = new JTextField(String.valueOf(((float) (int) (v2 * 100)) / 100f));// 計算
+		JTextField showIdealweight = new JTextField(String.valueOf(((int) (v2 * 100)) / 100f));// 計算
 		showIdealweight.setFont(font);
 		showIdealweight.setEditable(false);
 		showIdealweight.setBounds(200 + X, 200 + Y, 200, 64);
